Rename files for consistent naming convention
