+++ packages/cli/commands/fetchImage.ts
