+++ packages/cli/index.ts
@@ -2,8 +2,8 @@
 import { bold } from "cli-color"
 import { Command } from "commander"
 import { conductor } from "./commands/conductor"
-import { devSgs } from "./commands/dev-sgs"
-import { fetchImage } from "./commands/fetch-image"
+import { devSgs } from "./commands/devSgs"
+import { fetchImage } from "./commands/fetchImage"
 import { messageProcessing } from "./commands/message-processing"
 import { status } from "./commands/status"
 import { version } from "./package.json"
